Move 'LoginScreen' to screens directory.
