+++ examples/rn-todo-list/frontend/app/AppWrapperSync.tsx
@@ -22,7 +22,7 @@ import {OpenRealmBehaviorType} from 'realm';
 import {AppProvider, RealmProvider, UserProvider} from '@realm/react';
 
 import {AppSync} from './AppSync';
-import {LoginScreen} from './components/LoginScreen';
+import {LoginScreen} from './screens/LoginScreen';
 import {Task} from './models/Task';
 import {colors} from './styles/colors';
 import {schemas} from './models';
