+++ examples/rn-todo-list/frontend/app/styles/button.ts
@@ -17,6 +17,7 @@
 ////////////////////////////////////////////////////////////////////////////
 
 import {StyleSheet} from 'react-native';
+
 import {colors} from './colors';
 
 export const buttonStyles: StyleSheet.NamedStyles<any> = {
