+++ examples/rn-todo-list/frontend/app/styles/shadows.ts
@@ -17,6 +17,7 @@
 ////////////////////////////////////////////////////////////////////////////
 
 import {Platform} from 'react-native';
+
 import {colors} from './colors';
 
 export const shadows = Platform.select({
